requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
black==26.1.0
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
crispy-bootstrap5==2026.3
cryptography==46.0.5
decorator==5.2.1
defusedxml==0.7.1
Django==6.0.3
django-bootstrap==0.2.4
django-crispy-forms==2.6
django-extensions==4.1
django-sslserver==0.22
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
frozenlist==1.8.0
gunicorn==25.1.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
mac-vendor-lookup==0.1.15
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
multidict==6.7.1
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.9.2
port-scanner==0.1.4
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
python-nmap==0.7.1
pytokens==0.4.1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scapy==2.7.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.4
traitlets==5.14.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.6
yarg==0.1.9
yarl==1.23.0
